split the factory token into separate actor and implementation crates
This keeps the actor-specific code (mainly the invoke() function) away from the main implementation, so we can import it into other packages (like integration tests) without the invoke function causing conflicts at linking time.
