+++ Cargo.toml
@@ -17,7 +17,8 @@ members = [
     "testing/fil_token_integration/actors/basic_transfer_actor",
     "testing/fil_token_integration/actors/frc46_test_actor",
     "testing/fil_token_integration/actors/frcxx_test_actor",
-    "token_examples/factory_token"
+    "token_examples/factory_token",
+    "token_examples/factory_token/token_impl"
 ]
 
 [patch.crates-io]
