+++ token_examples/factory_token/Cargo.toml
@@ -15,6 +15,7 @@ fvm_shared = { version = "2.0.0-alpha.2" }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_tuple = { version = "0.5.0" }
 thiserror = { version = "1.0.31" }
+token_impl = { path = "token_impl" }
 
 [dev-dependencies]
 cid = { version = "0.8.5", default-features = false }
