+++ token_examples/factory_token/token_impl/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "token_impl"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+cid = { version = "0.8.5", default-features = false }
+frc42_dispatch = { path = "../../../frc42_dispatch" }
+frc46_token = { path = "../../../frc46_token" }
+fvm_actor_utils = { version = "0.1.0", path = "../../../fvm_actor_utils" }
+fvm_ipld_blockstore = "0.1.1"
+fvm_ipld_encoding = "0.2.2"
+fvm_sdk = { version = "2.0.0-alpha.2" }
+fvm_shared = { version = "2.0.0-alpha.2" }
+serde = { version = "1.0.136", features = ["derive"] }
+serde_tuple = { version = "0.5.0" }
+thiserror = { version = "1.0.31" }
+
